Refactor engine files.
