+++ src/engine/behavior.ts
