+++ src/engine/creature.ts
