+++ src/ui/components/expedition-screen.tsx
@@ -1,12 +1,12 @@
 /* eslint-disable max-len */
+import { Action, AttackAction, MoveByAction } from 'engine/actions'
+import { World } from 'engine/world'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import { useRecoilState, useRecoilValue, useResetRecoilState, useSetRecoilState } from 'recoil'
 import { useKeyHandler } from 'ui/hooks/use-key-handler'
 import { endTurn, expeditionState, isExpeditionComplete } from 'ui/state/expedition'
 import { gameState, pause, unpause } from 'ui/state/game'
 import { playerState } from 'ui/state/player'
-import { Action, AttackAction, MoveByAction } from 'world/actions'
-import { World } from 'world/world'
 
 import { ScreenName } from './app'
 import { LogPanel } from './log-panel'
